+++ src/models/Gasto.js
@@ -18,7 +18,7 @@ Gasto.init({
         allowNull: false
     },
     cuantia: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.FLOAT(11, 2),
         allowNull: false,
         validate: {
             min: 0,
